added alex's test flight feat
